Codex daemon review: apply findings; hardened round-trip verifies pixels
Second Codex adversarial review (context/impl/impl-review-findings-daemon.md,
3 P0 / 6 P1 / 6 P2 / 2 P3) on the daemon spine. Applied:
- P0 YapPacket signed-length overflow -> unsigned + bounded read.
- P0 tick reentrancy -> snapshot pages + defer page deletion (mReap) so a
  disconnect during nested GLib pumping can't free a page under the stack.
- P0 per-proxy multi-page: std::map<YapProxy*,Page> + per-page ProxySink routing.
- P1 shm ownership: adapter allocates buffers; daemon attaches-only, never
  IPC_RMIDs; adapter cleans up.
- P1 ordered GoannaRenderPage teardown (stop/remove-listener/clear-container).
- P2 dimension validation, openUrl failure -> LoadStopped, paint dedup
  (maybePaint once per load, not 60 Hz).
- P3 adapter verifies the shared buffer content (751756 non-white px).

Round-trip re-verified: ROUND-TRIP PASS, exit 0, with framebuffer content check.
Design-level items (non-blocking sends, returnBuffer double-buffering,
setWindowSize) dispositioned to T-016 in the findings file.

+++ render/browserserver/Yap/YapPacket.cpp
@@ -352,7 +352,10 @@ void YapPacket::operator>>(char*& val)
         g_return_if_fail(false);
     }
 
-    int16_t strLen = 0;
+    // Jihad hardening (Codex P0): use an UNSIGNED length so a wire value like
+    // 0xffff can't become negative, pass the bounds check, and blow up malloc/
+    // memcpy as a huge size_t. Bound the read against the packet length.
+    uint16_t strLen = 0;
 
     uint8_t* pSrc = m_buffer + m_currReadPos;
     uint8_t* pDst = (uint8_t*)(&strLen);
@@ -361,8 +364,8 @@ void YapPacket::operator>>(char*& val)
     pDst[1] = pSrc[0];
     m_currReadPos += 2;
 
-    g_return_if_fail((m_currReadPos + strLen) <= m_readTotalLen);
-    val = (char*) malloc(strLen + 1);
+    g_return_if_fail((size_t)(m_currReadPos + strLen) <= (size_t)m_readTotalLen);
+    val = (char*) malloc((size_t)strLen + 1);
 
     memcpy(val, m_buffer + m_currReadPos, strLen);
     val[strLen] = 0;
